changed back to useing the full brms model object
the full brms model object is used in the corrsponding vignette to illustrate model convergence and predictive power so it is retained in full despite being quite a large object.

That same RMR model vignette needs to be updated to the new code that currently resides in a script file.